+++ src/snowflake/cli/api/console/__init__.py
@@ -1,0 +1,3 @@
+from snowflake.cli.api.console.console import cli_console
+
+__all__ = ("cli_console",)