+++ RELEASE-NOTES.md
@@ -4,6 +4,8 @@
 
 ## New additions
 * Added ability to specify scope of the `object list` command with the `--in <scope_type> <scope_name>` option.
+* Introduced `snowflake.cli.api.console.cli_console` object with helper methods for intermediate output.
+
 ## Fixes and improvements